The julia exponent is now capable of being an arbitrary real number.
TODO:
-There is a discontinuity for at (x,0) for x<0 when the exponent is a
non integer. I'm not sure why this is. Maybe due to the math. I need to
figure out if it's supposed to be there or not.

-Using polar coords is naturally more computationally expoensive. The
checkbox in the julia control panel enabling the polar computatons is
not hooked up yet. The program needs to be modified to make this
selection matter.